{{cookiecutter.project_slug}}/{{cookiecutter.python_slug}}/requirements.txt
black
fastapi
isort
mypy
passlib[bcrypt]
pytest
pytest-xdist
pytest-cov
pylint
pydocstyle
pyjwt
python-multipart
redis
requests
uvicorn
